Introduced a helper module that defines all dummy classes
these dummy classes are used for testing. having them all at one comming
place allows to share them among different test and eases keeping track
which dummies are already there.
