+++ test/test_observer.rb
@@ -2,12 +2,6 @@ require "test_helper"
 
 class TestObserver < Minitest::Test
 
-  class ARandomClass < Observer
-  end
-
-  class BRandomClass < Observer
-  end
-
   def test_properly_inherited
     arc = ARandomClass.new
     assert_respond_to(arc, :handle_event)
@@ -17,16 +11,6 @@ class TestObserver < Minitest::Test
   end
 
   def test_implementing_abstract_methods_works
-    BRandomClass.class_eval do
-
-      def handle_event
-        puts "implemented 1"
-      end
-
-      def handle_event_with(message)
-        puts "implemented 2"
-      end
-    end
     arc = BRandomClass.new
     out1 = fetch_stdout {arc.handle_event}
     out2 = fetch_stdout {arc.handle_event_with(nil)}
